fix: complete French event descriptions and split MiniMap under 300 lines (#457)
- Translate the event-specific first sentence of 258 event.*.desc keys in
  fr.json that were left in English by the earlier translation pass
  (lawsuit 53, mafia 50, politics 53, union 51, weather 51). The shared
  boilerplate tail was already French and is preserved verbatim.
- Split src/ui/MiniMap.ts (306 lines) by extracting the canvas layer
  painting into src/ui/miniMapLayers.ts. MiniMap public API unchanged;
  drawNavGridOverlay stays on the class as a thin delegate.
